feat: add profileHeaderOption
